requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
openai==1.12.0
deepgram-sdk==4.1.0
python-multipart==0.0.9
aiohttp==3.9.3
httpx==0.26.0

# File upload dependencies
pytus>=0.2.1

# Testing dependencies
pytest==8.0.0
pytest-asyncio==0.23.5
rich==13.7.0
pytest-cov==4.1.0
